+++ changelog.md
@@ -1,13 +1,14 @@
 # pmxbot-glossary changelog
 
 **0.4.0**
-*(TBD)*
+*(Oct 21, 2014)*
 
 * Removed `!define` and replaced it with `!set`. !define overrides a pmxbot
 default, and the override stopped working for some unknown reason. Since there
 is not a clear way to "properly" override a default command, we'll use something
 new.
-
+* Removed author from standard output. 
+* Added !whowrote to get the author of an entry.
 
 * README fixes.